refactor of a few vocabularies components plus addition of Vocabularies table + form
