+++ src/components/dataset/table/DatasetTable.jsx
@@ -27,8 +27,6 @@ class DatasetTable extends React.Component {
   /**
    * HELPERS
    * - getDatasets
-   * - validate
-   * - isValid
   */
   getDatasets() {
     const { application } = this.props;
